requirements.txt
numpy~=1.23.3
sqlalchemy~=1.4.39
requests~=2.28.2
xmltodict~=0.13.0
google-api-python-client~=2.84.0